Stop JsRuntime from polluting global namespace
The previous implementation of JsRuntime was polluting the global namespace
by adding the JS script directly inside index.html. This commit removes that
and instead uses javascript modules throuhg IJsRuntime and IJSObjectReference.

This allows JS functions to be exported by the script and then imported using the
IJSObjectReference module inside the files where we need to use the JS function.